added logic og edit account

+++ components/data-table.tsx
@@ -2,6 +2,7 @@
 import * as React from 'react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
+import { useConfirm } from '@/hooks/use-confirm';
 import { Trash } from 'lucide-react';
 import {
     ColumnDef,
@@ -40,6 +41,10 @@ export function DataTable<TData, TValue>({
     onDelete,
     disabled,
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmDialog, confirm] = useConfirm(
+        'Are you sure?',
+        'You are about to perform a bulk delete.'
+    )
     const [sorting, setSorting] = React.useState<SortingState>([]);
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
         []
@@ -66,6 +71,7 @@ export function DataTable<TData, TValue>({
 
     return (
         <div>
+            <ConfirmDialog/>
             <div className="flex items-center py-4">
                 <Input
                     placeholder={`Filter ${filterKey}...`}
@@ -87,9 +93,12 @@ export function DataTable<TData, TValue>({
                         size="sm"
                         variant='outline'
                         className='ml-auto font-normal text-xs'
-                        onClick={() => {
-                            onDelete(table.getFilteredSelectedRowModel().rows)
-                            table.resetRowSelection()
+                        onClick={async () => {
+                            const ok = await confirm()
+                            if (ok) {
+                                onDelete(table.getFilteredSelectedRowModel().rows)
+                                table.resetRowSelection()
+                            }
                         }}
                     >
                         <Trash className='size-4 mr-2'/>
